+++ webapp/client/src/containers/NMDCHeader.js
@@ -91,7 +91,7 @@ const NMDCHeader = (props) => {
       ) : (
         <>
           <CHeaderNav className="nmdc-header-orcid nmdc-header-orcid-login">
-            <a href="/login" >
+            <a href="/oauth" >
               <img alt="OrcId login" src="https://orcid.org/assets/vectors/orcid.logo.icon.svg" className="nmdc-header-orcid-img mr-2" />
               OrcID Login
             </a>
