+++ webapp/client/src/routes.js
@@ -4,7 +4,6 @@ const Home = React.lazy(() => import('./pipelines/Home'));
 const Tutorial = React.lazy(() => import('./pipelines/Tutorial'));
 const PublicProjectList = React.lazy(() => import('./common/UM/Public/ProjectList'));
 const PublicProjectPage = React.lazy(() => import('./pipelines/Common/Results/ProjectPage/Public'));
-const UserLogin = React.lazy(() => import('./common/UM/Common/ORCIDLoginPage'));
 const tempLogin = React.lazy(() => import('./common/UM/User/Login'));
 // const UserActivate = React.lazy(() => import('./common/UM/User/Activate'));
 // const UserRegister = React.lazy(() => import('./common/UM/User/Register'));
@@ -18,7 +17,6 @@ const routes = [
   { path: '/tutorial', name: 'Home', component: Tutorial },
   { path: '/public/projectlist', name: 'ProjectList', component: PublicProjectList },
   { path: '/public/project', name: 'ProjectPage', component: PublicProjectPage },
-  { path: '/login', exact: true, name: 'Login', component: UserLogin },
   { path: '/nmdcedgeadminlogin', exact: true, name: 'tempLogin', component: tempLogin },
   // { path: '/activate', exact: true, name: 'Activate', component: UserActivate },
   // { path: '/register', exact: true, name: 'Register', component: UserRegister },
